:children_crossing: Reorder schema fields according to spec file
The fields are now displayed in the same order as they are put in the spec.
For this to work you should merge the spec into one file and insert the line
indexes yourself. See details in commit.

Before, fields were shown in alphabetic order - there may be a simplier way
to achieve the original order (when reading the file), but `x-index` is
already used to build search index for `lunr`.